refactor: extract app initialization logic into custom hook
Move the app initialization logic from App.tsx to a reusable custom hook, useAppInitialization, to improve code maintainability and separation of concerns. This change reduces code duplication and makes the initialization logic more modular.
